requirements.txt
# PRSL → prtextstyle 変換ツール 必要なパッケージ
# Python 3.8+ で動作

# 必須パッケージ
pillow>=9.0.0
numpy>=1.20.0

# 推奨パッケージ（高速化のため）
scipy>=1.7.0
